Move up ternary button creation to `HitObjectComposer` level
